# Project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/purchaseorders": {
            "get": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Get all purchase orders",
                "operationId": "24456f04271cbaae655249211564e3a8",
                "responses": {
                    "200": {
                        "description": "List of purchase orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "supplier_id": {
                                                "type": "integer"
                                            },
                                            "order_number": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "total_amount": {
                                                "type": "number"
                                            },
                                            "ordered_at": {
                                                "type": "string"
                                            },
                                            "received_at": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "created_by": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Create purchase order with items",
                "operationId": "c72025cf65d88e0cca2a00fbd6891c31",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "supplier_id",
                                    "order_number",
                                    "total_amount",
                                    "ordered_at",
                                    "created_by",
                                    "items"
                                ],
                                "properties": {
                                    "supplier_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_number": {
                                        "type": "string",
                                        "example": "PO-001"
                                    },
                                    "total_amount": {
                                        "type": "number",
                                        "example": 100
                                    },
                                    "ordered_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-14"
                                    },
                                    "received_at": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "created_by": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "items": {
                                        "description": "List of products in this purchase order",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase order created successfully"
                    },
                    "403": {
                        "description": "No Permission"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/purchaseorders/{purchases}": {
            "get": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Get single purchase order",
                "operationId": "ef46a3ba11d015c2cecdc89cc5b284bb",
                "parameters": [
                    {
                        "name": "purchases",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase order found"
                    },
                    "403": {
                        "description": "No Permission"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Update purchase order",
                "operationId": "8d63a70e42bf7c6a9085e253db8e8827",
                "parameters": [
                    {
                        "name": "purchases",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "supplier_id": {
                                        "type": "integer"
                                    },
                                    "order_number": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "total_amount": {
                                        "type": "number"
                                    },
                                    "ordered_at": {
                                        "type": "string"
                                    },
                                    "received_at": {
                                        "type": "string"
                                    },
                                    "created_by": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "403": {
                        "description": "No Permission"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Delete purchase order",
                "operationId": "cea688615ac9d15d5b1f71711fe43c67",
                "parameters": [
                    {
                        "name": "purchases",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "403": {
                        "description": "No Permission"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/purchaseorders/{id}/submit": {
            "post": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Submit purchase order",
                "operationId": "9ff512eb2314a0bee72756ab33e45385",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submitted successfully"
                    },
                    "400": {
                        "description": "Invalid status transition"
                    },
                    "403": {
                        "description": "No Permission"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/purchaseorders/{id}/receive": {
            "post": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Mark purchase order as received",
                "operationId": "aa1fc7136703bfce085ace3e5be79e8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Received successfully"
                    },
                    "400": {
                        "description": "Invalid status transition"
                    },
                    "403": {
                        "description": "No Permission"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/purchaseorders/{id}/cancel": {
            "post": {
                "tags": [
                    "Purchase Orders"
                ],
                "summary": "Cancel purchase order",
                "operationId": "9a46b2b7a5cce5efe70b51f5e9a116cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cancelled successfully"
                    },
                    "400": {
                        "description": "Invalid status transition"
                    },
                    "403": {
                        "description": "No Permission"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ],
...
}